fix: Move Content to own prop named FollowListItems
